Initial commit
"""
Real-Time Edge Detection using OpenCV in Python | Canny edge detection method
+ cv2.ORB_create() addition to display points of interest over video stream.
Welcome to a feature matching tutorial with OpenCV and Python. 
Feature matching is going to be a slightly more impressive version of template matching,
where a perfect, or very close to perfect, match is required.

We start with the image that we're hoping to find, and then we can search for this image within another image
produced by a camera stream. The beauty here is that the image does not need to be the same lighting, angle, rotation...etc. 
The features just need to match up.
author: Peter Mankowski
May 2020
"""